+++ Changelog.md
@@ -3,3 +3,6 @@ Changes in version 0.4.32
 -------------------------
   * changed an internal structure of ak_random class
   * added a winrtl random generator
+  * improved key masking for Kuznechik block cipher
+  * counter mode for 128-bit block ciphers
+